+++ computer-test/src/main/java/com/baidu/hugegraph/computer/algorithm/community/trianglecount/TriangleCountTest.java
@@ -122,7 +122,7 @@ public class TriangleCountTest extends AlgorithmTestBase {
         @Override
         public Long value(
                com.baidu.hugegraph.computer.core.graph.vertex.Vertex vertex) {
-            Long value = super.value(vertex);
+            Long value = (Long) super.value(vertex);
             Long expected = EXPECTED_RESULTS.get(vertex.id());
             if (expected != null) {
                 Assert.assertEquals(expected, value);
